% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gR2-package.r, R/gR2.R
\docType{package}
\name{gR2}
\alias{gR2}
\alias{gR2-package}
\title{gR2.}
\usage{
gR2(x, y, z = NULL, K = NULL, cand.Ks = 1:4, inference = FALSE,
  conf.level = 0.95, method = c("general", "binorm"), nstart = 30,
  mc.cores = detectCores() - 1)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector of the same length as \code{x}.}

\item{z}{an optional numeric vector containing integer values that indicate line memberships under the supervised scenario. The length of \code{z} must be the same as that of \code{x} and \code{y}.}

\item{K}{an optional number of lines under the unsupervised scenario, when \code{z=NULL}.}

\item{cand.Ks}{\code{1:4} (default) or a numeric vector containing the candidate values of \code{K}, when \code{K=NULL} under the unsupervised scenario (\code{z=NULL}).}

\item{inference}{logical. Should statistical inference be performed? Default is \code{FALSE}. If \code{TRUE}, a confidence interval of level \code{conf.level} will be computed for the population generalized R square measure, and a p-value will be computed for a one-sided test (greater than) against the null hypothesis that the population generalized R square measure is equal to zero.}

\item{conf.level}{the confidence level for the returned confidence interval. Default is \code{0.95}.}

\item{method}{a character string indicating which asymptotic distribution of the sample generalized measure is to be used for the inference. It must be one of \code{"general"} (the general asymptotic distribution) or \code{"binorm"} (the asymptotic distribution under the binormal distribution). The default is \code{"general"}.}

\item{nstart}{the number of initial starts for the K-lines algorithm under the unsupervised scenario (\code{z=NULL}).}

\item{mc.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously. Must be at least one, and parallelization requires at least two cores. The default is the number of CPU cores minus one.}
}
\value{
\code{gR2} returns a list with the following components:

\item{estimate}{The sample generalized R square measure.}
\item{conf.level}{The confidence level for the returned confidence interval (if \code{inference} is \code{TRUE}).}
\item{conf.int}{A numeric vector with two elements indicating the lower and upper bounds of the confidence interval (if \code{inference} is \code{TRUE}).}
\item{p.val}{The p-value for testing against the null hypothesis that the population generalized R square measure is zero (if \code{inference} is \code{TRUE}).}
\item{K}{The number of line relationships (if \code{z} is \code{NULL}).}
\item{membership}{The inferred line memberships of the data points in \code{x} and \code{y} (if \code{z} is \code{NULL}).}
}
\description{
\code{gR2} is the main function that calculates the sample generalized R square measures, i.e., point estimates, for the supervised scenario and the unsupervised scenario with or without K specified. It also performs statistical inference for the population measures, i.e., parameters of interest.
}
\details{
\code{x}, \code{y} and \code{z} (if not \code{NULL}) must be numeric vectors of the same length. If \code{z} is not \code{NULL}, the supervised scenario is considered; otherwise, the unsupervised scenario is considered.

Under the supervised scenario, \code{K} will not be used.

Under the unsupervised scenario, if \code{is.null(K)} is \code{FALSE}, \code{K} must be a positive integer. If users are interested in the existence of a mixture of two linear relationships, set \code{K=2}. We do not recommend specifying \code{K} larger than \code{4} for interpretability consideration. If \code{is.null(K)} is \code{TRUE}, the function will automatically choose a \code{K} value from candidate values (default set to \code{{1,2,3,4}}) using the Akaike information criterion (AIC). The function will output two plots: (1) a scree plot showing how the within-cluster average squared perpendicular distances vary with \code{K} values, and (2) a plot showing how AIC changes with \code{K}. Users can decide whether the chosen \code{K} value is reasonable by checking these two plots.

If \code{inference} is \code{FALSE} (default), the function only outputs a point estimate, i.e., a sample generalized R square measure. Under the unsupervised scenario, the inferred line memberships by the K-lines clustering will also be output. If \code{inference} is \code{TRUE}, the function will additionally output a \code{conf.level}-level confidence interval of the population generalized R square measure, as well as a p-value for a one-sided test against the null hypothesis that the population generalized R square is equal to zero.
}
\examples{
# generate data from a bivariate normal mixture model
library(mvtnorm)
library(parallel)
n = 200 # sample size
K = 2 # number of components (lines)
p_s = c(0.5, 0.5) # proportions of components
mu_s = list(c(0,-2), c(0,2)) # mean vectors
Sigma_s = list(rbind(c(1,0.8),c(0.8,1)), rbind(c(1,0.8),c(0.8,1))) # covariance matrices
z = sample(1:K, size=n, prob=p_s, replace=TRUE) # line memberships
data = matrix(0, nrow=n, ncol=2)
for (i in 1:K) {
  idx = which(z==i)
  data[idx,] = rmvnorm(n=length(idx), mean=mu_s[[i]], sigma=Sigma_s[[i]])
}
x = data[,1]
y = data[,2]

# supervised sample generalized R square
gR2(x, y, z) # without inference
gR2(x, y, z, inference=TRUE) # with inference

# unsupervised sample generalized R square
#gR2(x, y, K=2, mc.cores=2) # with K specified, without inference
#gR2(x, y, inference=TRUE, mc.cores=2) # without K specified, with inference

}
\references{
Li, J.J., Tong, X., and Bickel, P.J. (2018). Generalized R2 Measures for a Mixture of Bivariate Linear Dependences. arXiv.
}
\author{
Jingyi Jessica Li, \email{jli@stat.ucla.edu}
}
