\name{graphical.AdjRsqr}
\alias{graphical.AdjRsqr}
\title{Graphical adjusted R$^2$}
\usage{
  graphical.AdjRsqr(c = 1, R2c = rep(0, length(c)))
}
\arguments{
  \item{c}{Adjustment exponent}

  \item{R2c}{Vector of c-hypotheses}
}
\value{
  NULL
}
\description{
  Diagram for graphically adjusting sample R$^2$ values
}

