\name{R2.baseline}
\alias{R2.baseline}
\title{Baseline R$^2$}
\usage{
  R2.baseline(p, n, c, R2c = 0)
}
\arguments{
  \item{p}{Number of parameters}

  \item{n}{Number of samples}

  \item{c}{Adjustment exponent}

  \item{R2c}{c-hypothesis}
}
\value{
  Expected sample R$^2$
}
\description{
  Expected R$^2$ under a c-hypothesis
}

