% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gR2.R
\name{gR2}
\alias{gR2}
\title{Estimate population generalized R square measures}
\usage{
gR2(x, y, z = NULL, K = NULL, cand.Ks = 1:4, inference = FALSE,
  conf.level = 0.95, method = c("general", "binorm"), nstart = 30,
  mc.cores = detectCores() - 1)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector of the same length as \code{x}.}

\item{z}{an optional numeric vector containing integer values that indicate line memberships under the supervised scenario. The length of \code{z} must be the same as that of \code{x} and \code{y}.}

\item{K}{an optional number of lines under the unsupervised scenario, when \code{z=NULL}.}

\item{cand.Ks}{\code{1:4} (default) or a numeric vector containing the candidate values of \code{K}, when \code{K=NULL} under the unsupervised scenario (\code{z=NULL}).}

\item{inference}{logical. Should statistical inference be performed? Default is \code{FALSE}. If \code{TRUE}, a confidence interval of level \code{conf.level} will be computed for the population generalized R square measure, and a p-value will be computed for a one-sided test (greater than) against the null hypothesis that the population generalized R square measure is equal to zero.}

\item{conf.level}{the confidence level for the returned confidence interval. Default is \code{0.95}.}

\item{method}{a character string indicating which asymptotic distribution of the sample generalized measure is to be used for the inference. It must be one of \code{"general"} (the general asymptotic distribution) or \code{"binorm"} (the asymptotic distribution under the binormal distribution). The default is \code{"general"}.}

\item{nstart}{the number of initial starts for the K-lines algorithm under the unsupervised scenario (\code{z=NULL}).}

\item{mc.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously. Must be at least one, and parallelization requires at least two cores. The default is the number of CPU cores minus one.}
}
\value{
dd
}
\description{
\code{gR2} is the main function that calculates the sample generalized R square measures, i.e., point estimates, for the supervised scenario and the unsupervised scenario with or without K specified. It also performs statistical inference for the population measures, i.e., parameters of interest.
}
\details{
\code{x}, \code{y} and \code{z} (if not \code{NULL}) must be numeric vectors of the same length. If \code{z} is not \code{NULL}, the supervised scenario is considered; otherwise, the unsupervised scenario is considered.

Under the supervised scenario, \code{K} will not be used.

Under the unsupervised scenario, if \code{is.null(K)} is \code{FALSE}, \code{K} must be a positive integer. If users are interested in the existence of a mixture of two linear relationships, \code{K=2}. We do not recommend specifying \code{K} larger than four for interpretability consideration. If \code{is.null(K)} is \code{TRUE}, the function will automatically choose a \code{K} value from candidate values (default set to \code{{1,2,3,4}}) using the Akaike information criterion (AIC). The function will output two plots: (1) a scree plot showing how the within-cluster average squared perpendicular distances vary with \code{K} values, and (2) a plot showing how AIC changes with \code{K}. Users can decide whether the chosen \code{K} value is reasonable by checking these two plots.

If
}
