% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_gR2.R
\name{gR2}
\alias{gR2}
\title{gR2}
\usage{
gR2(
  x,
  y,
  z = NULL,
  K = NULL,
  cand.Ks = 1:4,
  nstart = 30,
  mc.cores = NULL,
  regressionMethod = "MA",
  verbose = TRUE,
  inference = FALSE,
  conf.level = 0.95,
  gR2.pop = 0,
  alternative = "greater",
  method = "general",
  details = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}

\item{z}{A vector of integers that represents the line membership of all the data points. Must be of the same length as \code{x} and \code{y}. Default is \code{NULL}.}

\item{K}{Number of lines in the unspecified scenario. Default is \code{NULL}.}

\item{cand.Ks}{A vector of positive integers that represents the candidate K’s in the unspecified scenario. Default is \code{1:4}.}

\item{nstart}{Number of initializations for the K-lines algorithm in the unspecified scenario. Default is \code{30}.}

\item{mc.cores}{Number of cores to use in the unspecified scenario. Default is \code{NULL}, which means all of the available cores are used.}

\item{regressionMethod}{Valid values are \code{‘MA’} and \code{‘LM’}. Indicates which regression method to use in the K-lines algorithm - major axis regression or linear regression. Default is \code{‘MA’}.}

\item{verbose}{Logical. If \code{TRUE}, then messages are printed and a graph is produced in the unspecified scenario (K not chosen). Default is \code{TRUE}.}

\item{inference}{Logical. If \code{TRUE}, then the function calculates a confidence interval for the population gR2 of confidence level \code{conf.level}, as well as a p-value of the hypothesis test where the null hypothesis is that the population gR2 is equal to \code{gR2.pop}. Default is \code{FALSE}.}

\item{conf.level}{The confidence level of the confidence interval. Default is \code{0.95}.}

\item{gR2.pop}{The population gR2 in the null hypothesis of the hypothesis test. Must be between \code{0} and \code{1}. Default is \code{0}.}

\item{alternative}{Valid values are \code{‘two.sided’}, \code{‘less’}, and \code{‘greater’}. Indicates the type of the alternative hypothesis in the hypothesis test.}

\item{method}{Valid values are \code{‘general’} and \code{‘binorm’}. Indicates which asymptotic distribution of the sample gR2 to use for inference. Default is \code{‘general’}.}

\item{details}{Logical. If \code{TRUE}, then detailed information about each line (R squared and its corresponding p-value) is outputted. Only available in the unspecified scenario. Default is \code{FALSE}.}
}
\value{
\code{gR2} returns a list consisting of one or more of the following items:
\item{estimate}{The sample gR2.}
\item{conf.level}{The confidence level of the confidence interval (if \code{inference} is \code{TRUE}).}
\item{conf.int}{The confidence interval for the population gR2 (if \code{inference} is \code{TRUE}).}
\item{p.val}{The p-value of the hypothesis test where the null hypothesis is that the population gR2 is equal to \code{gR2.pop} and the alternative hypothesis is that the population gR2 is not equal to, less than, or greater than \code{gR2.pop} depending on \code{alternative} (if \code{inference} is \code{TRUE}).}
\item{K}{The number of lines in the unspecified scenario, either chosen by the user or chosen from \code{cand.Ks} by the \code{gR2} function.}
\item{membership}{The inferred line membership of all the data points in the unspecified scenario.}
\item{perLineInfo}{A matrix with three columns: lineIndex, R2, and pValue. Each row corresponds to a line. Total number of rows is \code{K}, the number of lines chosen.}
}
\description{
\code{gR2} calculates the sample gR2 under the specified scenario, the unspecified scenario (K chosen), and the unspecified scenario (K not chosen). It also provides an option to perform statistical inference on the population gR2.
}
\details{
The arguments that require user input are \code{x} and \code{y}, which must be numeric vectors of the same length.

There are three broad types of scenarios: the specified scenario, the unspecified scenario (K chosen), and the unspecified scenario (K not chosen). The specified scenario is considered when \code{z} is provided; the unspecified scenario (K chosen) is considered when \code{z} is not provided but \code{K} is provided; and the unspecified scenario (K not chosen) is considered when neither \code{z} or \code{K} is provided.

In the unspecified scenario (K chosen), we recommend that users set \code{K} to be less than or equal to \code{4} for interpretability.

In the unspecified scenario (K not chosen), the \code{gR2} function automatically chooses a \code{K} value from \code{cand.Ks} using the Akaike information criterion (AIC). Two plots are outputted unless \code{verbose} is set to \code{FALSE}: (1) a scree plot that shows how average squared perpendicular/vertical distance changes with the candidate \code{K}, and (2), a plot that shows how AIC changes with the candidate \code{K}. Users can decide whether the \code{K} value chosen by the \code{gR2} function is reasonable by checking these two plots.
}
\references{
Li, J.J., Tong, X., and Bickel, P.J. (2019). Generalized R2 Measures for a Mixture of Bivariate Linear Dependences. arXiv.
}
\author{
Jingyi Jessica Li, \email{jli@stat.ucla.edu}

Heather J Zhou, \email{heatherjzhou@ucla.edu}
}
